requirements.txt
streamlit
pandas
requests
python-dotenv
supabase
beautifulsoup4
lxml
geopy
playwright
firebase-admin
apify-client
openpyxl
pycryptodome
wmi
customtkinter
